Replace legacy login activity with new firebase UI logic
