+++ app/src/main/AndroidManifest.xml
@@ -21,7 +21,8 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme"
         tools:ignore="GoogleAppIndexingWarning">
-        <activity android:name=".ui.login.FirebaseAuthActivity"></activity>
+        <activity android:name=".ui.login.LoginActivity"
+            android:theme="@style/AppTheme.LoginScreen"></activity>
         <activity
             android:name=".ui.splash.SplashActivity"
             android:configChanges="orientation|screenSize"
