+++ app/src/main/res/layout/fragment_login_landing.xml
@@ -1,64 +1,41 @@
 <?xml version="1.0" encoding="utf-8"?>
-<FrameLayout
+<ScrollView
     xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:id="@+id/login_scrollview"
     android:layout_width="match_parent"
-    android:layout_height="match_parent" >
+    android:layout_height="match_parent"
+    android:layout_gravity="center_horizontal"
+    android:paddingTop="@dimen/activity_vertical_margin"
+    android:background="?attr/colorPrimary">
 
-    <ScrollView
-        android:id="@+id/login_scrollview"
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:id="@+id/register_layout"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:paddingLeft="@dimen/activity_horizontal_margin"
-        android:paddingRight="@dimen/activity_horizontal_margin"
-        android:paddingTop="@dimen/activity_vertical_margin"
-        android:paddingBottom="@dimen/activity_vertical_margin"
-        android:background="?attr/colorPrimary"
-        tools:context=".ui.login.LoginActivity">
+        android:layout_height="wrap_content"
+        android:paddingHorizontal="@dimen/activity_horizontal_margin">
 
-        <LinearLayout
-            android:id="@+id/register_layout"
+        <ImageView
+            android:id="@+id/logo"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintTop_toTopOf="parent"
+            android:layout_width="74dp"
+            android:layout_height="74dp"
+            android:contentDescription="@string/video_logo"
+            android:src="@drawable/video_logo_splash"/>
+
+        <Button
+            android:id="@+id/sign_in_button"
+            app:layout_constraintTop_toBottomOf="@id/logo"
+            app:layout_constraintStart_toStartOf="parent"
+            android:layout_marginTop="@dimen/username_margin_top"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:paddingLeft="@dimen/activity_horizontal_margin"
-            android:paddingRight="@dimen/activity_horizontal_margin"
-            android:paddingTop="@dimen/activity_vertical_margin"
-            android:paddingBottom="@dimen/activity_vertical_margin"
-            android:orientation="vertical" >
+            android:textAppearance="?android:attr/textAppearanceMedium"
+            android:textColor="@android:color/white"
+            android:background="?attr/colorPrimaryDark"
+            android:text="@string/log_in"/>
 
-            <ImageView
-                android:layout_width="74dp"
-                android:layout_height="74dp"
-                android:layout_gravity="center_horizontal"
-                android:contentDescription="@string/video_logo"
-                android:src="@drawable/video_logo_splash"/>
-
-            <com.google.android.gms.common.SignInButton
-                android:id="@+id/google_sign_in_button"
-                android:layout_marginTop="@dimen/username_margin_top"
-                android:layout_marginLeft="@dimen/activity_horizontal_margin"
-                android:layout_marginStart="@dimen/activity_horizontal_margin"
-                android:layout_marginRight="@dimen/activity_horizontal_margin"
-                android:layout_marginEnd="@dimen/activity_horizontal_margin"
-                android:layout_marginBottom="@dimen/activity_horizontal_margin"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"/>
-
-            <Button
-                android:id="@+id/email_sign_in_button"
-                android:layout_marginTop="@dimen/login_components_margin_top"
-                android:layout_marginLeft="@dimen/activity_horizontal_margin"
-                android:layout_marginStart="@dimen/activity_horizontal_margin"
-                android:layout_marginRight="@dimen/activity_horizontal_margin"
-                android:layout_marginEnd="@dimen/activity_horizontal_margin"
-                android:layout_marginBottom="@dimen/activity_horizontal_margin"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:textAppearance="?android:attr/textAppearanceMedium"
-                android:textColor="@android:color/white"
-                android:background="?attr/colorPrimaryDark"
-                android:text="@string/sign_in_with_existing_account"/>
-        </LinearLayout>
-    </ScrollView>
-
-</FrameLayout>
+    </androidx.constraintlayout.widget.ConstraintLayout>
+</ScrollView>
